requirements.txt
flask==2.1.0
werkzeug==2.0.3
boto3==1.28.0
python-dotenv==0.19.0